Port src/marks/{rect,bar} from JS to TS (stage 1)
Combine each .js + .d.ts pair into a single .ts file with no behaviour
changes. Types/JSDoc come from the .d.ts; runtime from the .js.
Cross-mark import bar.ts -> ./rect.js resolves cleanly.
